requirements.txt
certifi==2020.12.5
chardet==4.0.0
ci-info==0.2.0
click==7.1.2
configobj==5.0.6
configparser==5.0.2
decorator==4.4.2
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
future==0.18.2
httplib2==0.19.0
idna==2.10
isodate==0.6.0
lxml==4.6.2
networkx==2.5
nibabel==3.2.1
nipype==1.6.0
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pathlib==1.0.1
Pillow==8.1.0
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.18.9
pyparsing==2.4.7
PyQt5==5.15.3
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2021.1
pyxnat==1.4
rdflib==5.0.0
requests==2.25.1
scipy==1.5.4
simplejson==3.17.2
six==1.15.0
traits==6.2.0
urllib3==1.26.3
